requirements.txt
tensorflow 
tensorflow-addons
tqdm
pysoundfile
pydub
scipy
librosa
flask
